+++ decidim-assemblies/app/commands/decidim/assemblies/admin/notify_role_assigned_assembly.rb
@@ -5,11 +5,11 @@ module Decidim
     module Admin
       # A command to notify users when a role is assigned for a participatory process
       class NotifyRoleAssignedAssembly < Rectify::Command
-        def send_notification user
+        def send_notification (user)
             Decidim::EventsManager.publish(
               event: "decidim.events.assembly.role_assigned",
               event_class: Decidim::AssemblyRoleAssignedEvent,
-              resource: @assembly,
+              resource: form.current_participatory_space,
               affected_users: [user],
               extra: {
                 role: form.role