+++ decidim-assemblies/spec/commands/update_assembly_admin_spec.rb
@@ -20,6 +20,15 @@ module Decidim::Assemblies
     end
     let(:current_user) { create(:user, :admin, :confirmed) }
     let(:invalid) { false }
+    let(:user_notification) do
+      {
+        event: "decidim.events.assembly.role_assigned",
+        event_class: AssemblyRoleAssignedEvent,
+        resource: assembly,
+        affected_users: [user],
+        extra: { role: kind_of(String) }
+      }
+    end
 
     context "when the form is not valid" do
       let(:invalid) { true }
@@ -42,6 +51,12 @@ module Decidim::Assemblies
         expect do
           subject.call
         end.to change { user_role.reload && user_role.role }.from("admin").to(new_role)
+      end
+
+      it "sends a notification to the user with the updated role" do
+        expect(Decidim::EventsManager).to receive(:publish).with(user_notification)
+
+        subject.call
       end
 
       it "broadcasts ok" do