+++ decidim-conferences/app/commands/decidim/conferences/admin/notify_role_assigned_conference.rb
@@ -5,16 +5,16 @@ module Decidim
     module Admin
     # A command to notify users when a role is assigned for a Conference
       class NotifyRoleAssignedConference < Rectify::Command
-        def send_notification user
-            Decidim::EventsManager.publish(
+        def send_notification (user)
+          Decidim::EventsManager.publish(
             event: "decidim.events.conferences.role_assigned",
             event_class: Decidim::Conferences::ConferenceRoleAssignedEvent,
-            resource: @conference,
+            resource: form.current_participatory_space,
             affected_users: [user],
             extra: {
                 role: form.role
             }
-            )
+          )
         end
       end
     end