+++ decidim-participatory_processes/app/commands/decidim/participatory_processes/admin/notify_role_assigned_participatory_process.rb
@@ -5,11 +5,11 @@ module Decidim
     module Admin
       # A command to notify users when a role is assigned for a participatory process
       class NotifyRoleAssignedParticipatoryProcess < Rectify::Command
-        def send_notification user
+        def send_notification (user)
             Decidim::EventsManager.publish(
               event: "decidim.events.participatory_process.role_assigned",
               event_class: Decidim::ParticipatoryProcessRoleAssignedEvent,
-              resource: @participatory_process,
+              resource: form.current_participatory_space,
               affected_users: [user],
               extra: {
                 role: form.role